Relocate KB-related components
