+++ components/kb/AppKbDisplay.js
