+++ components/kb/AppKbView.js
